requirements.txt
streamlit
openai
pandas
matplotlib
pyPDF2
pillow
fcsparser